requirements.txt
gunicorn==20.0.4
opencv_python==4.8.0.74
tensorflow== 2.13.0rc0
algorithmia==1.3.0
scikit-video==1.1.11
tf_slim==1.1.0
PyYaml==5.3.1
matplotlib==3.7.1
opencv-python~=4.8.0.74
numpy~=1.25.0
Pillow~=10.0.0